# swagger_apisix.yaml
openapi: 3.0.0
info:
  title: AI Rate Limiter API
  version: 1.0.0
  description: API documentation for AI Rate Limiter service
servers:
  - url: http://localhost:8501
    description: Local development server
paths:
...
  /providers:
    get:
      tags:
        - Provider
      summary: List all providers
      description: List all AI providers. Optionally filter by queue id.
      parameters:
        - in: query
          name: queue_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of providers
          content:
            application/json:
              schema:
                type: object
                properties:   
                  success:        
                    type: boolean
                    example: true
                  data:         
                    type: array
                    items:
                      $ref: '#/components/schemas/Provider'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /provider/create:
    post:
      tags:
        - Provider
      summary: Create an AI provider
      description: Create a new AI provider.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queue_id:
                  type: string
                  example: "84cc-43ec-bc4f-0000-000000000009"
                provider_name:
                  type: string
                  example: "OpenAI"
                api_key:
                  type: string
                  example: "sk-proj-1234567890"
                limit:
                  type: integer
                  example: 1000
                time_window:
                  type: integer
                  example: 3600
                config:
                  type: object
                  properties:
                    model:
                      type: string
                      example: "gpt-4o"
                    azure_api_version:
                      type: string
                    endpoint:
                      type: string
              required:
                - model
                - api_key
                - limit
                - time_window
      responses:
        '201':
          description: Provider added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Provider added successfully"
                  success:
                    type: boolean
                    example: true
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Queue not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /provider/update/{provider_id}:
    patch:
      tags:
        - Provider
      summary: Update provider by id
      description: Update an existing AI provider by its ID and store the changes in the database.
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queue_id:
                  type: string
                  description: The queue to which this provider belongs
                  example: "84cc-43ec-bc4f-0000-000000000009"
                provider_name:
                  type: string
                  example: "OpenAI"
                api_key:
                  type: string
                  example: "sk-proj-1234567890"
                limit:
                  type: integer
                  example: 1000
                time_window:
                  type: integer
                  example: 3600
                config:
                  type: object
                  properties:
                    model:
                      type: string
                      example: "gpt-4o"
                    azure_api_version:
                      type: string
                      example: "2024-05-01-preview"
                    endpoint:
                      type: string
                      example: "https://api.openai.com"
              required:
                - queue_id
                - model
                - api_key
                - limit
                - time_window
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Provider updated successfully"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /provider/delete/{provider_id}:
    delete:
      tags:
        - Provider
      summary: Delete a provider
      description: Delete an AI provider by its ID.
      parameters:
        - in: path
          name: provider_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Provider deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Provider key 24e583ef-eb8a-4c04-93df-06e606716d56 deleted successfully"
                  success:
                    type: boolean
                    example: true
        '404':
          description: Provider not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      example:
        message: "string"
        success: false
...
    Provider:
      type: object
      properties:
        provider_id:
          type: string
          example: "24e583ef-eb8a-4c04-93df-06e606716d56"
        queue_id:
          type: string
          example: "-84cc-43ec-bc4f-000000000009"
        provider_name:
          type: string
          example: "OpenAI"
        provider_type:
          type: string
          enum:
            - openai
            - azure
            - anthropic
            - deepseek
            - claude
        api_key:
          type: string
          example: "sk-proj-1234567890"
        limit:
          type: integer
          example: 1000
        time_window:
          type: integer
          example: 3600
        config:
          type: object
          properties:
            model:
              type: string
              example: "gpt-4o"
            azure_api_version:
              type: string
              example: "2024-05-01-preview"
            endpoint:
              type: string
              example: "https://api.openai.com"
        created_at:
          type: string
          format: date-time
          example: "2024-06-01T12:34:56Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-06-01T12:34:56Z"
      required:
        - provider_name
        - api_key
        - limit
        - time_window
        - config
...